Spli consistent-dev behavior in a diff file
